Refactor ecosystem support into integrations
Move Transformers discovery/config loading and Diffusers pipeline support under dedicated integration packages. Centralize shared Hugging Face weight loading, preserve legacy private module aliases and public signatures, and add compatibility coverage for imports and monkeypatch injection points.


Copilot-Session: 39ff6824-d35f-4d3b-8f5f-043a7119a100
