+++ src/mobius/integrations/gguf/_builder.py
@@ -337,14 +337,11 @@ def build_from_gguf(
             keep_quantized=keep_quantized,
         )
 
-    from mobius._builder import (
+    from mobius._registry import registry
+    from mobius.integrations._builder import (
         build_from_module,
         resolve_dtype,
     )
-    from mobius._config_resolver import (
-        _default_task_for_model,
-    )
-    from mobius._registry import registry
     from mobius.integrations.gguf._config_mapping import (
         GGUF_ARCH_TO_MODEL_TYPE,
         gguf_to_config,
@@ -352,6 +349,9 @@ def build_from_gguf(
     from mobius.integrations.gguf._reader import GGUFModel
     from mobius.integrations.gguf._tensor_processors import (
         process_tensors,
+    )
+    from mobius.integrations.transformers import (
+        _default_task_for_model,
     )
 
     if static_cache and task is not None:
