+++ src/mobius/integrations/nemo/_builder.py
@@ -57,11 +57,11 @@ def build_from_nemo(
     """
     import dataclasses
 
-    from mobius._builder import build_from_module, resolve_dtype
-    from mobius._config_resolver import _default_task_for_model
     from mobius._registry import registry
+    from mobius.integrations._builder import build_from_module, resolve_dtype
     from mobius.integrations.nemo._config_mapping import nemo_to_config
     from mobius.integrations.nemo._reader import NeMoArchive
+    from mobius.integrations.transformers import _default_task_for_model
 
     archive = NeMoArchive(nemo_path, revision=revision)
     logger.info("Loaded NeMo archive: %s (target=%s)", archive.path, archive.target)
