+++ src/mobius/tasks/_qwen_image.py
@@ -9,8 +9,8 @@ from typing import ClassVar
 
 import onnx_ir as ir
 
-from mobius._diffusers_configs import QwenImageConfig
 from mobius._model_package import ModelPackage
+from mobius.integrations.diffusers._configs import QwenImageConfig
 from mobius.tasks._base import ModelTask, _make_graph, _make_model
