+++ src/mobius/tasks/_qwen_image_vae.py
@@ -14,8 +14,8 @@ from typing import ClassVar
 
 import onnx_ir as ir
 
-from mobius._diffusers_configs import QwenImageVAEConfig
 from mobius._model_package import ModelPackage
+from mobius.integrations.diffusers._configs import QwenImageVAEConfig
 from mobius.tasks._base import ComponentSpec, ModelTask, _make_graph, _make_model
