+++ src/mobius/tasks/_vae.py
@@ -14,8 +14,8 @@ from typing import ClassVar
 
 import onnx_ir as ir
 
-from mobius._diffusers_configs import VAEConfig
 from mobius._model_package import ModelPackage
+from mobius.integrations.diffusers._configs import VAEConfig
 from mobius.tasks._base import ComponentSpec, ModelTask, _make_graph, _make_model
