+++ src/mobius/tasks/_video_denoising.py
@@ -14,8 +14,8 @@ from typing import ClassVar
 
 import onnx_ir as ir
 
-from mobius._diffusers_configs import CogVideoXConfig
 from mobius._model_package import ModelPackage
+from mobius.integrations.diffusers._configs import CogVideoXConfig
 from mobius.tasks._base import ModelTask, _make_graph, _make_model
